pyproject.toml
[project]
name = "platform-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.134.0",
    "crewai-tools>=0.0.1",
    "langchain-openai>=0.3.27",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
